[INLONG-9948][Agent] The instance class has many code similarities, abstract a common class
